+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * `plot.Inspector` can be scrolled using arrow keys or mouse wheel (#36)
 * New `plot.Systems` for inspecting ECS systems (#36)
+* New `plot.Resources` for inspecting ECS resources (#36)
 
 ## [[v0.3.0]](https://github.com/mlange-42/arche-pixel/compare/v0.2.0...v0.3.0)